rewrite uniqnum test to compare entries rather than lists
The uniqnum.t test previously compared the resulting lists when using
two different uniqueness implementations.  This gave a working pass/fail
check, but made diagnostics hard.

Rewrite the test to instead run the routines against all of the
previously entries when iterating over all of the source numbers.  This
makes it easier to report proper test names, as we keep the source code
that resulted in the numbers, as well as being able to directly compare
how the routines are treating individual values.  It also means that
test counts are consistent across machines.

Also simplify the argument parsing, adding the --diag option to fully
report the internal representations of the values we are trying to
compare.